spring security in journal endpoints

+++ src/main/java/net/engineeringdigest/journalApp/Service/JournalEntryService.java
@@ -27,7 +27,7 @@ public class JournalEntryService {
         User user= userService.findByUserName(userName);
         JournalEntry saved = journalEntryRepository.save(journalEntry);
         user.getJournalEntries().add(saved);
-        userService.saveEntry(user);
+        userService.saveUser(user);
 
     }
 
@@ -45,10 +45,22 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user= userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public Boolean deleteById(ObjectId id, String userName){
+        Boolean removed = false;
+        try {
+            User user= userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting entry.",e);
+        }
+
+        return removed;
     }
+
 }
